PUBDEV-8385: remove kwargs parameter, improve H2OPlotResult structure and move it to its own package
