+++ h2o-py/h2o/model/extensions/scoring_history.py
@@ -1,5 +1,5 @@
 from h2o.exceptions import H2OValueError
-from h2o.plot_result import H2OPlotResult
+from h2o.plot.plot_result import H2OPlotResult
 from h2o.utils.ext_dependencies import get_matplotlib_pyplot
 from h2o.utils.shared_utils import can_use_pandas
 from h2o.utils.typechecks import assert_is_type
@@ -30,7 +30,7 @@ class ScoringHistory:
     def _validate_timestep(self, timestep):
         return timestep
     
-    def scoring_history_plot(self, timestep, metric, server=False, save_plot_path=None, **savefig):
+    def scoring_history_plot(self, timestep, metric, server=False, save_plot_path=None):
         plt = get_matplotlib_pyplot(server)
         if plt is None: return
         
@@ -72,7 +72,7 @@ class ScoringHistory:
             plt.ylim(ylim)
             plt.plot(scoring_history[timestep], scoring_history[training_metric])
         if save_plot_path is not None:
-            plt.savefig(fname=save_plot_path, fig=fig, **savefig)    
+            plt.savefig(fname=save_plot_path, fig=fig)    
         if not server:
             plt.show()
         return H2OPlotResult(figure=fig)    
@@ -111,7 +111,7 @@ class ScoringHistoryGLM(ScoringHistory):
         # for others, validation is done in the plot function below
     )
     
-    def scoring_history_plot(self, timestep, metric, server=False, save_plot_path=None, **savefig):
+    def scoring_history_plot(self, timestep, metric, server=False, save_plot_path=None):
         plt = get_matplotlib_pyplot(server)
         if plt is None: return
         
@@ -146,7 +146,7 @@ class ScoringHistoryGLM(ScoringHistory):
         style = "b-" if len(scoring_history[timestep]) > 1 else "bx"
         plt.plot(scoring_history[timestep], scoring_history[metric], style)
         if save_plot_path is not None:
-            plt.savefig(fname=save_plot_path, fig=fig, **savefig)
+            plt.savefig(fname=save_plot_path, fig=fig)
         if not server:
             plt.show()
         return H2OPlotResult(figure=fig)
