+++ h2o-py/h2o/plot/__init__.py
@@ -1,0 +1,3 @@
+from .plot_result import H2OPlotResult
+
+__all__ = ["H2OPlotResult"]
