+++ h2o-py/tests/testdir_misc/pyunit_plot_roc_pr_curve.py
@@ -55,9 +55,9 @@ def plot_test():
  
         
      # Test no plot parameter
-    (fprs, tprs) = perf_test.plot(type="roc", server=True, plot=False).tuple()
+    (fprs, tprs) = perf_test.plot(type="roc", server=True, plot=False)
     assert len(fprs) == len(tprs), "Expected fprs and tprs to have the same shape but they are not."
-    (recalls, precisions) = perf_test.plot(type="pr", server=True, plot=False).tuple()
+    (recalls, precisions) = perf_test.plot(type="pr", server=True, plot=False)
     assert len(recalls) == len(precisions), "Expected recall and precision to have the same shape but they are not."
